requirements.txt
streamlit
streamlit-extras
pandas
numpy
yfinance
plotly
matplotlib
pandas_ta
google-generativeai
FinMind